+++ dwave/optimization/_model.pxd
