+++ dwave/optimization/_model.pyi
