+++ dwave/optimization/_model.pyx
