+++ dwave/optimization/mathematical.py
@@ -16,7 +16,7 @@ import collections
 import functools
 import typing
 
-from dwave.optimization._graph import ArraySymbol
+from dwave.optimization._model import ArraySymbol
 from dwave.optimization.symbols import (
     Add,
     And,
